bugfix: broadcastGetBlock failed when syncblock
